backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent101-backend"
version = "0.1.0"
description = "Backend API for Agent101"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Agent101 Team"}
]
keywords = ["api", "backend", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "a2a-sdk>=0.3.0",
    "a2a-server>=0.6.1",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "google-adk>=1.18.0",
    "web3>=6.15.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    # ag-ui-adk: AG-UI Protocol adapter for Google ADK agents
    # Allows ADK agents to communicate with the frontend using AG-UI Protocol
    "ag-ui-adk>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

# Removed project.scripts.start - use uvicorn directly or create a proper entry point

[tool.setuptools]
packages = ["packages", "agents"]
# Note: packages/blockchain/dex is included in packages

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long, handled by black
    "E402",  # module level import not at top of file (intentional in some cases)
    "E722",  # do not use bare except (intentional in some error handling)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"agents/multichain_liquidity/__main__.py" = ["E402"]
"agents/balance/__main__.py" = ["E402"]
"agents/sentiment/__main__.py" = ["E402"]
"agents/trading/__main__.py" = ["E402"]
"packages/blockchain/hedera/balance/balance_client.py" = ["E722"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests", "packages"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=packages",
    "--cov=agents",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "-v",
]

[tool.coverage.run]
source = ["packages", "agents"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

